Rewire analytics, decisions and assistant from fixtures to real data
The Buying insights ledger and every module that read alongside it were
built on hashed/fixture arithmetic even after the foundation commit made
purchase_order, deal and the catalogue real. This finishes the job:

analytics: BuyAnalyticsEngine reads supplier name/country straight off
each PoRow (already real, written by the importer or recordAward) and
pulls vendorCode/priceIndex/qualityPpm/creditDays/leadTimeDays from
history.Suppliers, nullable rather than defaulted to 0 when the panel
doesn't have them. ProcurementAnalyticsService.recordAward resolves
branch name/region from history.Catalogue, promised days from the
supplier's known lead time (else observed lead from
history.PurchaseHistory) plus history.Reference's origin transit, sets
product_id/store_id, and evicts history's caches after commit. Categories
now reflects products.category instead of a hand-mapped pipe/tube/valves
scheme. Deleted the whole internal/fixtures seed-JSON package.

decisions: HistoryEngine's SELL_MONTH_ORDER is derived from
AatlasClock.today() (twelve months ending this month) instead of a fixed
Oct-Sep array; item short names come from history.Catalogue instead of
the deleted ProductNames seed reader. Fixed a latent NPE in
HistoryService's ?sort=margin (marginPct is null for a sell deal with no
cost on file - a real, schema-allowed state since V22 - and
comparingDouble would unbox it).

assistant: AssistantCatalog (a classpath seed shared by every tenant) is
gone; AssistantService now resolves products/stores/regions through
history.Catalogue, so "Tell me about Dallas" and "copper" questions
answer over this tenant's own real rows. Every intent that needs sales
or purchase history a tenant hasn't uploaded yet returns an "Upload
sales history to unlock this" answer instead of a hollow zero, and
which-supplier/hold-vs-sell try every real candidate the question could
mean before giving up, guarding against bulk's still-fixture-based
readers reporting non-priceable.

Gates green: mvnw test (146 tests) and mvnw verify -Dit.test=
AnalyticsProcurementIT,DecisionsHistoryIT,AssistantIT (28 tests) against
Docker/Testcontainers Postgres.

+++ src/main/java/com/aatlas/analytics/internal/ledger/SupplierRow.java
@@ -1,8 +1,16 @@
 package com.aatlas.analytics.internal.ledger;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-/** Ports {@code procurement.ts}'s {@code SupplierRow} - the Buying insights scorecard row. */
+/**
+ * Ports {@code procurement.ts}'s {@code SupplierRow} - the Buying insights scorecard row.
+ *
+ * <p>{@code priceIndex}, {@code avgLeadDays}, {@code qualityPpm} and {@code creditDays} are
+ * boxed: null when the supplier panel does not carry the fact (see {@link SupplierFacts}) and
+ * there is no window of received orders to observe it from either - "not assessed", never a
+ * placeholder.
+ */
 public record SupplierRow(
         String id,
         String name,
@@ -15,11 +23,11 @@ public record SupplierRow(
         double saved,
         double leaked,
         double avgLanded,
-        double priceIndex,
+        BigDecimal priceIndex,
         double onTimePct,
-        int avgLeadDays,
-        int qualityPpm,
-        int creditDays,
+        Integer avgLeadDays,
+        Integer qualityPpm,
+        Integer creditDays,
         String risk,
         List<Double> series,
         int colorIndex) {
